+++ EXAMPLES/clustering/clustering_14.cpp
@@ -49,7 +49,7 @@ void clustering_14_example() {
 
 	// nbCluster contains the numbers of clusters to be tested.
 	// Here we assume that there are 2, 3 or 4 clusters.
-	vector<int64_t> nbCluster;
+	vector<int> nbCluster;
 	nbCluster.push_back (2);
 	//~ nbCluster.push_back (3);
 	//~ nbCluster.push_back (4);
@@ -110,7 +110,7 @@ void clustering_14_example() {
 		// print out parameters
 		param->edit();
 		// print out criterion values
-		for (int64_t iCriterion = 0; iCriterion < cInput->getCriterionName().size(); iCriterion++)
+		for (int iCriterion = 0; iCriterion < cInput->getCriterionName().size(); iCriterion++)
 			cMOutput->getCriterionOutput (cInput->getCriterionName (iCriterion)).editTypeAndValue (std::cout);
 	}
 	cout << "-----------------------------------------------------------------------" << endl;
