+++ EXAMPLES/discriminant_analysis/discriminant_analysis_1.cpp
@@ -93,8 +93,8 @@ void discriminant_analysis_1_example() {
     XEM::LearnModelOutput * lMOutput = NULL;
 
     // TODO: next 3 lines should not be needed (always use vector ?)
-    int64_t* vLabel_p = lInput->getKnownLabelDescription()->getLabel()->getTabLabel();
-    vector<int64_t> vLabel;
+    int* vLabel_p = lInput->getKnownLabelDescription()->getLabel()->getTabLabel();
+    vector<int> vLabel;
     for (int j=0; j<lInput->getNbSample(); j++) vLabel.push_back(vLabel_p[j]);
 
     if ( lOutput->atLeastOneEstimationNoError() ) {
@@ -130,7 +130,7 @@ void discriminant_analysis_1_example() {
 
         std::cout << "CV Error Rate:" << ( 1-lMOutput->getCVLabel()->getLabel()->getErrorRate ( vLabel ) ) *100 << "%" << std::endl;
 
-        int64_t ** classif = lMOutput->getCVLabel()->getLabel()->getClassificationTab ( vLabel, lMOutput->getNbCluster() );
+        int ** classif = lMOutput->getCVLabel()->getLabel()->getClassificationTab ( vLabel, lMOutput->getNbCluster() );
         std::cout << "CV Classification Tab:" << std::endl;
         for ( int i=0; i<lMOutput->getNbCluster(); i++ ) {
             std::cout << " | " ;
